new test_forecast function
It takes the shelve files with cell properties (output of properties_cells),
computes the history back (up to one hour or when the cell disappears) and
extrapolates according to the model chosen for up to 1 hour (unless the
area becomes smaller then zero). It then plots the contours for +20 min
+40 min and +1h on top of the coalition2 rgb for all the cells (existing for
at least 1 timestep before the current).