NEW version of plot_coalition2 integrated/faster
New version of plot_coalition2 and the supporting scripts and input
file that are more efficient (optimized reading of shelve files)
and are all combined and integrated within plot_coalition. If the
area chosen is ccs4 and the corresponding switches in input_coalition2
are turned on (in_msg.properties_cells and in_msg.plot_forecast) the
rgb with the area forecast is produced.
Still to correct: check for area local or europe in plot_coalition2.py
(check_input_file function).