New script for validation
new scripts that allows to choose multiple cells (and t_start and t_end)
for which it builds the history, extrapolates, builds history forward and
evaluate the accuracy of the forecast. It plots (if set to true) scatter
plot of errors (position, dx, dy, area, 108), corresponding 6bins histograms
and SouthNorth-EastWest position error plot.